libs/tts/edge-tts/requirements.txt
edge-tts
pyzmq
